fix: move cors to file app
