+++ server.js
@@ -1,16 +1,7 @@
 const app = require('./app');
 const { sequelize } = require('./models');
 const PORT = process.env.PORT || 3000;
-const cors = require('cors');
 
-const corsConfig = {
-  origin: '*',
-  credentials: true,
-  methods: ['GET', 'POST', 'PUT', 'DELETE'],
-};
-
-app.options('*', cors(corsConfig));
-app.use(cors(corsConfig));
 
 sequelize.authenticate()
   .then(() => {
